refactor(axios): refactored the axios factory and allowed for a custom alias to be used on request hook
