+++ lib/index.js
@@ -4,7 +4,7 @@ const debug = require('debug')('hapi:plugins:good-tracer');
 const NodeCache = require('node-cache');
 const { v4: uuidv4 } = require('uuid');
 const { get, set, isNaN } = require('lodash');
-const axiosFactory = require('./axiosFactory');
+const { factory: axiosFactory } = require('./axios');
 
 const NAME = 'goodTracer';
 
@@ -163,8 +163,8 @@ exports.register = (server, options) => {
     }
 };
 
-exports.addAxiosRoutePreHook = {
-    assign: 'axios',
+exports.addAxiosRoutePreHookFactory = (assign = 'axios') => ({
+    assign,
     method: (request) => {
         const factory = get(request, ['server', 'plugins', NAME, 'axiosFactory']);
         const { traceUUIDHeader, traceDepthHeader, config } = get(request, ['server', 'plugins', NAME, 'axiosConfig'], {});
@@ -181,4 +181,6 @@ exports.addAxiosRoutePreHook = {
 
         return factory(axiosConfig);
     }
-};
+});
+
+exports.addAxiosRoutePreHook = exports.addAxiosRoutePreHookFactory();
