+++ spec/hooks.spec.js
@@ -1,10 +1,12 @@
 const { set } = require('lodash');
-const { addAxiosRoutePreHook } = require('../lib');
-const axiosFactory = require('../lib/axiosFactory');
+const { addAxiosRoutePreHook, addAxiosRoutePreHookFactory } = require('../lib');
+const { factory } = require('../lib/axios');
 
 describe('addAxiosRoutePreHook', () => {
-    it('assign to axios', () => {
+    it('should assign to axios', () => {
         expect(addAxiosRoutePreHook.assign).toBe('axios');
+        expect(addAxiosRoutePreHook.method).toBeDefined();
+        expect(addAxiosRoutePreHook.method).toBeInstanceOf(Function);
     });
 
     it('return an axios client with default headers', () => {
@@ -18,7 +20,7 @@ describe('addAxiosRoutePreHook', () => {
             }
         };
         const request = {};
-        set(request, 'server.plugins.goodTracer.axiosFactory', axiosFactory);
+        set(request, 'server.plugins.goodTracer.axiosFactory', factory);
         set(request, 'server.plugins.goodTracer.axiosConfig', {
             traceUUIDHeader,
             traceDepthHeader,
@@ -35,5 +37,21 @@ describe('addAxiosRoutePreHook', () => {
             depth: 1234,
             'user-agent': 'test-mctestings'
         });
+    });
+});
+
+describe('addAxiosRoutePreHookFactory', () => {
+    it('should assign to axios by default', () => {
+        const result = addAxiosRoutePreHookFactory();
+        expect(result.assign).toBe('axios');
+        expect(result.method).toBeDefined();
+        expect(result.method).toBeInstanceOf(Function);
+    });
+
+    it('should assign to the passed in string', () => {
+        const result = addAxiosRoutePreHookFactory('yolo');
+        expect(result.assign).toBe('yolo');
+        expect(result.method).toBeDefined();
+        expect(result.method).toBeInstanceOf(Function);
     });
 });
